# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Users, Documentation",
        "description": "Esta API é utilizada para gerenciar um sistema de acessos em geral, oferecendo funcionalidades para autenticação de usuários, gerenciamento de contas, e administração de privilégios. Ela suporta três níveis de acesso: Administrador, Atendente e Cliente.\n  - Administradores podem acessar e modificar dados de qualquer usuário.\n  - Atendentes podem gerenciar contas, como update, deletar usuarios e ver uma lista completa de dados de cada usuario e acessar algumas funcionalidades restritas.\n  - Clientes têm acesso limitado às suas próprias contas.\n\n **Autenticação**\n - A API utiliza autenticação baseada em tokens **(Bearer Token via Sanctum)**. Todos os endpoints que requerem autenticação estão devidamente marcados com a configuração de segurança. O objetivo principal desta documentação é ajudar desenvolvedores a integrar suas aplicações com o sistema, fornecendo detalhes sobre requisições, respostas, e exemplos práticos de uso.\n\n  **Principais funcionalidades**:\n  - Registro, login e logout de usuários.\n  - Gestão de privilégios para controle de acesso.\n  - Operações CRUD em usuários (somente admins e atendentes).",
        "contact": {
            "email": "oliveirarafael22@outlook.com.br"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://api-users-rafael.up.railway.app/api/v1/",
...
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "Administração"
                ],
                "summary": "Realiza uma Collection de todos os usuarios ",
                "description": "Somente admins e atendentes podem Realizar essa consulta",
                "operationId": "23679253d16f34d2972249751e3631dd",
                "responses": {
                    "200": {
                        "description": "OK, Dados recuperados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "2"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "edward@teste.com"
                                                    },
                                                    "privilege_id": {
                                                        "type": "integer",
                                                        "example": "3"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Você não tem permissão para acessar este recurso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Somente administradores poderao ver todas as contas do sistema como um collection",
                                            "type": "string",
                                            "example": "You are not authorized to access"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado. Faça login para continuar.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/users": {
            "post": {
                "tags": [
                    "Administração"
                ],
                "summary": "Acessa o Store , porem o acesso é somente a admins",
                "description": "O metodo Store ainda nao possui funcionalidade, porem através de policies o seu acesso é somente para admins e atendentes",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "responses": {
                    "403": {
                        "description": "Você não tem permissão para acessar este recurso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Somente administradores e atendentes terão acesso a essa função",
                                            "type": "string",
                                            "example": "You are not authorized to access"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado. Faça login para continuar.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
        "/user/{id}": {
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Excluir um Usuario,Somente admins e atendentes podem excluir contas de terceiros",
                "description": "Deleta o Usuario, Somente o administrador e Atendente poderá excluir outras contas, caso contrario o propio usuarios podera excluir sua conta.",
                "operationId": "a56b12a3694a926699892056cdcf8a9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuario a ser excluido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User removed success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Retornara uma mensagem de sucesso que o usuario foi exluido com sucesso",
                                            "type": "string",
                                            "example": "User removed success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Você não tem permissão para accessar este recurso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Somente administradores de sistema e atendentes podem excluir contas de terceiros",
                                            "type": "string",
                                            "example": "You are not authorized to access"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario nao encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Se o usuario nao existir no banco de dados ou se ja foi exluido",
                                            "type": "string",
                                            "example": "User Not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado. Faça login para continuar.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Adicione o token no cabeçalho Authorization no formato: Bearer <seu_token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}